Rename database-test to test-dashboard and fix dashboard routing conflict
