+++ app/(dashboard)/dashboard/page.tsx
@@ -1,286 +1,246 @@
 'use client';
 
-import { Button } from '@/components/ui/button';
-import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
-import {
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-  CardFooter
-} from '@/components/ui/card';
-import { customerPortalAction } from '@/lib/payments/actions';
-import { useActionState } from 'react';
-import { GroupWithMembers, Person } from '@/lib/db/schema';
-import { removeUser, inviteUser } from '@/app/(login)/actions';
-import useSWR from 'swr';
-import { Suspense } from 'react';
-import { Input } from '@/components/ui/input';
-import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group';
-import { Label } from '@/components/ui/label';
-import { Loader2, PlusCircle } from 'lucide-react';
+import React from 'react';
+import UniversalCard from '@/components/ui/UniversalCard';
+import UniversalButton from '@/components/ui/UniversalButton';
+import { 
+  BarChart3, 
+  Users, 
+  Calendar, 
+  ClipboardList, 
+  ArrowUpRight, 
+  Clock, 
+  Bell, 
+  ChevronRight, 
+  CalendarPlus,
+  UserPlus,
+  ClipboardPlus
+} from 'lucide-react';
 
-type ActionState = {
-  error?: string;
-  success?: boolean;
-  message?: string;
-};
+const mockStats = [
+  { title: 'Total Players', value: 42, icon: <Users size={24} className="text-gold-500" />, change: '+3', period: 'from last month' },
+  { title: 'Active Teams', value: 5, icon: <Users size={24} className="text-gold-500" />, change: '+1', period: 'from last month' },
+  { title: 'Upcoming Sessions', value: 12, icon: <Calendar size={24} className="text-gold-500" />, change: '+5', period: 'from last week' },
+  { title: 'Total Drills', value: 86, icon: <ClipboardList size={24} className="text-gold-500" />, change: '+8', period: 'from last month' },
+];
 
-const fetcher = (url: string) => fetch(url).then((res) => res.json());
+const mockUpcomingSessions = [
+  { id: 1, title: 'Team Practice - Varsity', date: '2025-07-11', time: '16:00', location: 'Main Court', players: 12 },
+  { id: 2, title: 'Shooting Drills - Advanced', date: '2025-07-12', time: '14:30', location: 'Training Room', players: 8 },
+  { id: 3, title: 'Junior Development', date: '2025-07-13', time: '10:00', location: 'Secondary Court', players: 15 },
+];
 
-function SubscriptionSkeleton() {
-  return (
-    <Card className="mb-8 h-[140px]">
-      <CardHeader>
-        <CardTitle>Team Subscription</CardTitle>
-      </CardHeader>
-    </Card>
-  );
-}
+const mockRecentActivities = [
+  { id: 1, type: 'player_added', message: 'New player Michael Johnson added to Varsity team', time: '2 hours ago' },
+  { id: 2, type: 'session_completed', message: 'Shooting Drills session completed with 8 players', time: '1 day ago' },
+  { id: 3, type: 'observation_added', message: 'New observation added for player Sarah Williams', time: '2 days ago' },
+  { id: 4, type: 'drill_created', message: 'New defensive drill "Zone Pressure" created', time: '3 days ago' },
+  { id: 5, type: 'team_updated', message: 'Junior team roster updated with 3 new players', time: '4 days ago' },
+];
 
-function ManageSubscription() {
-  const { data: teamData } = useSWR<GroupWithMembers>('/api/team', fetcher);
+const mockPlayerPerformance = [
+  { name: 'James Wilson', points: 18.5, rebounds: 7.2, assists: 4.3, trend: 'up' },
+  { name: 'Sarah Williams', points: 15.2, rebounds: 3.1, assists: 6.8, trend: 'up' },
+  { name: 'Marcus Johnson', points: 12.7, rebounds: 8.5, assists: 2.1, trend: 'down' },
+  { name: 'Emma Davis', points: 14.3, rebounds: 4.2, assists: 5.6, trend: 'stable' },
+];
 
-  return (
-    <Card className="mb-8">
-      <CardHeader>
-        <CardTitle>Team Subscription</CardTitle>
-      </CardHeader>
-      <CardContent>
-        <div className="space-y-4">
-          <div className="flex flex-col sm:flex-row justify-between items-start sm:items-center">
-            <div className="mb-4 sm:mb-0">
-              <p className="font-medium">
-                Current Plan: {teamData?.planName || 'Free'}
-              </p>
-              <p className="text-sm text-muted-foreground">
-                {teamData?.subscriptionStatus === 'active'
-                  ? 'Billed monthly'
-                  : teamData?.subscriptionStatus === 'trialing'
-                  ? 'Trial period'
-                  : 'No active subscription'}
-              </p>
-            </div>
-            <form action={customerPortalAction}>
-              <Button type="submit" variant="outline">
-                Manage Subscription
-              </Button>
-            </form>
-          </div>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
-
-function TeamMembersSkeleton() {
-  return (
-    <Card className="mb-8 h-[140px]">
-      <CardHeader>
-        <CardTitle>Team Members</CardTitle>
-      </CardHeader>
-      <CardContent>
-        <div className="animate-pulse space-y-4 mt-1">
-          <div className="flex items-center space-x-4">
-            <div className="size-8 rounded-full bg-gray-200"></div>
-            <div className="space-y-2">
-              <div className="h-4 w-32 bg-gray-200 rounded"></div>
-              <div className="h-3 w-14 bg-gray-200 rounded"></div>
-            </div>
-          </div>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
-
-function TeamMembers() {
-  const { data: teamData } = useSWR<GroupWithMembers>('/api/team', fetcher);
-  const [removeState, removeAction, isRemovePending] = useActionState<
-    ActionState,
-    FormData
-  >(removeUser, {});
-
-  const getUserDisplayName = (user: Pick<Person, 'id' | 'displayName' | 'email'>) => {
-    return user.displayName || user.email || 'Unknown User';
-  };
-
-  if (!teamData?.members?.length) {
-    return (
-      <Card className="mb-8">
-        <CardHeader>
-          <CardTitle>Team Members</CardTitle>
-        </CardHeader>
-        <CardContent>
-          <p className="text-muted-foreground">No team members yet.</p>
-        </CardContent>
-      </Card>
-    );
-  }
-
-  return (
-    <Card className="mb-8">
-      <CardHeader>
-        <CardTitle>Team Members</CardTitle>
-      </CardHeader>
-      <CardContent>
-        <ul className="space-y-4">
-          {teamData.members.map((member) => (
-            <li key={member.person.id} className="flex items-center justify-between">
-              <div className="flex items-center space-x-4">
-                <Avatar>
-                  {/* 
-                    This app doesn't save profile images, but here
-                    is how you'd show them:
-
-                    <AvatarImage
-                      src={member.user.image || ''}
-                      alt={getUserDisplayName(member.user)}
-                    />
-                  */}
-                  <AvatarFallback>
-                    {getUserDisplayName(member.person)
-                      .split(' ')
-                      .map((n) => n[0])
-                      .join('')}
-                  </AvatarFallback>
-                </Avatar>
-                <div>
-                  <p className="font-medium">
-                    {getUserDisplayName(member.person)}
-                  </p>
-                  <p className="text-sm text-muted-foreground capitalize">
-                    {member.role}
-                  </p>
-                </div>
-              </div>
-              <form action={removeAction}>
-                <input type="hidden" name="userId" value={member.person.id} />
-                <Button
-                  type="submit"
-                  variant="outline"
-                  size="sm"
-                  disabled={isRemovePending}
-                >
-                  {isRemovePending ? 'Removing...' : 'Remove'}
-                </Button>
-              </form>
-            </li>
-          ))}
-        </ul>
-        {removeState?.error && (
-          <p className="text-red-500 mt-4">{removeState.error}</p>
-        )}
-      </CardContent>
-    </Card>
-  );
-}
-
-function InviteTeamMemberSkeleton() {
-  return (
-    <Card className="h-[260px]">
-      <CardHeader>
-        <CardTitle>Invite Team Member</CardTitle>
-      </CardHeader>
-    </Card>
-  );
-}
-
-function InviteTeamMember() {
-  const { data: user } = useSWR<Person>('/api/user', fetcher);
-  // This check for 'owner' might need to be updated based on your specific logic
-  // for who can invite. For now, we'll check the superadmin flag or the contextual role.
-  const canInvite = user?.isSuperadmin; // Or check for a specific role in teamData
-  const [inviteState, inviteAction, isInvitePending] = useActionState<
-    ActionState,
-    FormData
-  >(inviteUser, {});
-
-  return (
-    <Card>
-      <CardHeader>
-        <CardTitle>Invite Team Member</CardTitle>
-      </CardHeader>
-      <CardContent>
-        <form action={inviteAction} className="space-y-4">
-          <div>
-            <Label htmlFor="email" className="mb-2">
-              Email
-            </Label>
-            <Input
-              id="email"
-              name="email"
-              type="email"
-              placeholder="Enter email"
-              required
-              disabled={!canInvite}
-            />
-          </div>
-          <div>
-            <Label>Role</Label>
-            <RadioGroup
-              defaultValue="member"
-              name="role"
-              className="flex space-x-4"
-              disabled={!canInvite}
-            >
-              <div className="flex items-center space-x-2 mt-2">
-                <RadioGroupItem value="member" id="member" />
-                <Label htmlFor="member">Member</Label>
-              </div>
-              <div className="flex items-center space-x-2 mt-2">
-                <RadioGroupItem value="owner" id="owner" />
-                <Label htmlFor="owner">Owner</Label>
-              </div>
-            </RadioGroup>
-          </div>
-          {inviteState?.error && (
-            <p className="text-red-500">{inviteState.error}</p>
-          )}
-          {inviteState?.success && (
-            <p className="text-green-500">{inviteState.success}</p>
-          )}
-          <Button
-            type="submit"
-            className="bg-orange-500 hover:bg-orange-600 text-white"
-            disabled={isInvitePending || !canInvite}
-          >
-            {isInvitePending ? (
-              <>
-                <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                Inviting...
-              </>
-            ) : (
-              <PlusCircle className="mr-2 h-4 w-4" />
-            )}
-            Invite
-          </Button>
-        </form>
-      </CardContent>
-      {!canInvite && (
-        <CardFooter>
-          <p className="text-sm text-muted-foreground">
-            You must be a team owner to invite new members.
-          </p>
-        </CardFooter>
-      )}
-    </Card>
-  );
-}
-
-export default function SettingsPage() {
+export default function Dashboard() {
   return (
     <section className="flex-1 p-4 lg:p-8">
-      <h1 className="text-lg lg:text-2xl font-medium mb-6">Team Settings</h1>
-      <Suspense fallback={<SubscriptionSkeleton />}>
-        <ManageSubscription />
-      </Suspense>
-      <Suspense fallback={<TeamMembersSkeleton />}>
-        <TeamMembers />
-      </Suspense>
-      <Suspense fallback={<InviteTeamMemberSkeleton />}>
-        <InviteTeamMember />
-      </Suspense>
+      <h1 className="text-lg lg:text-2xl font-medium mb-6">Dashboard</h1>
+      
+      {/* Stats Overview */}
+      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6 mb-8">
+        {mockStats.map((stat, index) => (
+          <UniversalCard.StatCard key={index} className="h-full">
+            <div className="flex items-start justify-between">
+              <div>
+                <p className="text-sm font-medium text-zinc-400">{stat.title}</p>
+                <h3 className="text-3xl font-bold mt-2 text-white">{stat.value}</h3>
+                <p className="flex items-center mt-1 text-xs text-gold-500">
+                  <ArrowUpRight size={14} className="mr-1" />
+                  {stat.change} {stat.period}
+                </p>
+              </div>
+              <div className="p-3 rounded-full bg-zinc-800">
+                {stat.icon}
+              </div>
+            </div>
+          </UniversalCard.StatCard>
+        ))}
+      </div>
+
+      <div className="grid grid-cols-1 lg:grid-cols-3 gap-6">
+        {/* Left Column - Upcoming Sessions */}
+        <div className="lg:col-span-2 space-y-6">
+          {/* Upcoming Sessions */}
+          <UniversalCard.Default
+            title="Upcoming Sessions"
+            subtitle="Next 7 days"
+            footer={
+              <div className="flex justify-between items-center w-full">
+                <span className="text-sm text-zinc-400">Showing {mockUpcomingSessions.length} of 12 sessions</span>
+                <UniversalButton.Secondary size="sm" rightIcon={<ChevronRight size={16} />}>
+                  View All
+                </UniversalButton.Secondary>
+              </div>
+            }
+          >
+            <div className="space-y-4">
+              {mockUpcomingSessions.map((session) => (
+                <div key={session.id} className="flex items-center p-3 rounded-lg bg-zinc-800/50 hover:bg-zinc-800 transition-colors">
+                  <div className="flex-shrink-0 w-12 h-12 bg-gold-500/20 rounded-lg flex items-center justify-center mr-4">
+                    <Calendar className="h-6 w-6 text-gold-500" />
+                  </div>
+                  <div className="flex-1 min-w-0">
+                    <p className="text-sm font-medium text-white truncate">{session.title}</p>
+                    <div className="flex items-center mt-1">
+                      <Clock size={14} className="text-zinc-400 mr-1" />
+                      <p className="text-xs text-zinc-400">
+                        {new Date(session.date).toLocaleDateString('en-US', { month: 'short', day: 'numeric' })} at {session.time}
+                      </p>
+                    </div>
+                  </div>
+                  <div className="ml-4">
+                    <div className="flex items-center">
+                      <Users size={14} className="text-zinc-400 mr-1" />
+                      <span className="text-xs text-zinc-400">{session.players} players</span>
+                    </div>
+                    <p className="text-xs text-zinc-500 mt-1">{session.location}</p>
+                  </div>
+                </div>
+              ))}
+            </div>
+          </UniversalCard.Default>
+
+          {/* Top Player Performance */}
+          <UniversalCard.Default
+            title="Top Player Performance"
+            subtitle="Last 30 days"
+          >
+            <div className="overflow-x-auto">
+              <table className="w-full">
+                <thead>
+                  <tr className="text-left text-xs text-zinc-500 border-b border-zinc-800">
+                    <th className="pb-2 font-medium">Player</th>
+                    <th className="pb-2 font-medium">PPG</th>
+                    <th className="pb-2 font-medium">RPG</th>
+                    <th className="pb-2 font-medium">APG</th>
+                    <th className="pb-2 font-medium">Trend</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {mockPlayerPerformance.map((player, index) => (
+                    <tr key={index} className="border-b border-zinc-800/50 hover:bg-zinc-800/30">
+                      <td className="py-3 pr-4">
+                        <p className="font-medium text-white">{player.name}</p>
+                      </td>
+                      <td className="py-3 pr-4 text-zinc-300">{player.points}</td>
+                      <td className="py-3 pr-4 text-zinc-300">{player.rebounds}</td>
+                      <td className="py-3 pr-4 text-zinc-300">{player.assists}</td>
+                      <td className="py-3">
+                        <span className={`inline-flex items-center px-2 py-0.5 rounded text-xs font-medium ${
+                          player.trend === 'up' ? 'bg-success-500/20 text-success-500' :
+                          player.trend === 'down' ? 'bg-danger-500/20 text-danger-500' :
+                          'bg-zinc-500/20 text-zinc-400'
+                        }`}>
+                          {player.trend === 'up' ? '↑ Up' : 
+                           player.trend === 'down' ? '↓ Down' : '→ Stable'}
+                        </span>
+                      </td>
+                    </tr>
+                  ))}
+                </tbody>
+              </table>
+            </div>
+          </UniversalCard.Default>
+        </div>
+
+        {/* Right Column - Activities and Quick Actions */}
+        <div className="space-y-6">
+          {/* Quick Actions */}
+          <UniversalCard.Default title="Quick Actions">
+            <div className="grid grid-cols-1 gap-3">
+              <UniversalButton.Primary 
+                leftIcon={<UserPlus size={16} />}
+                className="justify-start"
+              >
+                Add New Player
+              </UniversalButton.Primary>
+              
+              <UniversalButton.Primary 
+                leftIcon={<CalendarPlus size={16} />}
+                className="justify-start"
+              >
+                Schedule Session
+              </UniversalButton.Primary>
+              
+              <UniversalButton.Primary 
+                leftIcon={<ClipboardPlus size={16} />}
+                className="justify-start"
+              >
+                Create New Drill
+              </UniversalButton.Primary>
+            </div>
+          </UniversalCard.Default>
+
+          {/* Recent Activity */}
+          <UniversalCard.Default
+            title="Recent Activity"
+            subtitle="System updates and notifications"
+          >
+            <div className="space-y-4 max-h-[400px] overflow-y-auto pr-2">
+              {mockRecentActivities.map((activity) => (
+                <div key={activity.id} className="flex items-start p-3 rounded-lg bg-zinc-800/50 hover:bg-zinc-800 transition-colors">
+                  <div className="flex-shrink-0 w-8 h-8 bg-gold-500/20 rounded-full flex items-center justify-center mr-3">
+                    <Bell size={14} className="text-gold-500" />
+                  </div>
+                  <div>
+                    <p className="text-sm text-white">{activity.message}</p>
+                    <p className="text-xs text-zinc-500 mt-1">{activity.time}</p>
+                  </div>
+                </div>
+              ))}
+            </div>
+          </UniversalCard.Default>
+
+          {/* System Stats */}
+          <UniversalCard.Default title="System Health">
+            <div className="space-y-3">
+              <div>
+                <div className="flex justify-between mb-1">
+                  <span className="text-xs text-zinc-400">Storage</span>
+                  <span className="text-xs text-zinc-400">75%</span>
+                </div>
+                <div className="w-full bg-zinc-800 rounded-full h-2">
+                  <div className="bg-gold-500 h-2 rounded-full" style={{ width: '75%' }}></div>
+                </div>
+              </div>
+              
+              <div>
+                <div className="flex justify-between mb-1">
+                  <span className="text-xs text-zinc-400">User Accounts</span>
+                  <span className="text-xs text-zinc-400">42/50</span>
+                </div>
+                <div className="w-full bg-zinc-800 rounded-full h-2">
+                  <div className="bg-gold-500 h-2 rounded-full" style={{ width: '84%' }}></div>
+                </div>
+              </div>
+              
+              <div>
+                <div className="flex justify-between mb-1">
+                  <span className="text-xs text-zinc-400">Video Storage</span>
+                  <span className="text-xs text-zinc-400">40%</span>
+                </div>
+                <div className="w-full bg-zinc-800 rounded-full h-2">
+                  <div className="bg-gold-500 h-2 rounded-full" style={{ width: '40%' }}></div>
+                </div>
+              </div>
+            </div>
+          </UniversalCard.Default>
+        </div>
+      </div>
     </section>
   );
 }
