+++ middleware.ts
@@ -1,7 +1,8 @@
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
 import { signToken, verifyToken } from '@/lib/auth/session';
-import { refreshSession } from '@/lib/supabase/middleware';
+// Temporarily disable Supabase session refresh until the package is installed
+// import { refreshSession } from '@/lib/supabase/middleware';
 
 const protectedRoutes = ['/dashboard', '/admin'];
 const adminRoutes = ['/admin'];
@@ -20,10 +21,10 @@ export async function middleware(request: NextRequest) {
   }
 
   // Refresh Supabase auth session if needed
-  const supabaseRes = await refreshSession(request);
+  // const supabaseRes = await refreshSession(request);
 
   // Use the response from Supabase refresh (if any) as our base response
-  let res = supabaseRes ?? NextResponse.next();
+  let res = /* supabaseRes ?? */ NextResponse.next();
 
   if (sessionCookie) {
     try {
